requirements.txt
flask
gTTS
google-generativeai